Filter deleted threads in list immediately when deleted
When deleting a thread, we previously changed the content but left it
in the Pinia store. This worked correctly for the thread detail page,
but meant that the deleted thread also showed up in the thread list.
This commit makes it so that we remove the thread from the list as
soon as it's deleted by updating (and passing down from the API) the
`is_deleted` flag.