Add ArticleDates spec and update props-passing in tests
It looks like using `propsData` doesn't work anymore; we have to use
`props` now instead. Two files had been copied that were using
`propsData`, but they don't actually depend on the value to pass, so I
just didn't notice that it wasn't working.

PostForm was getting passed a prop it doesn't even take, so that was
just removed.

ReplyList does use its prop, but it doesn't cause an error to not pass
it in because it's just an ID we pass to store functions, and we mock
those functions return values but don't check that they're called. I
think this is appropriate, because the tests should be checking
external behaviour and don't need to couple themselves to how the
store functions are called; the typechecker can handle that.