Get prod preview working with Nuxt app
The /api/ prefix for the backend had to be moved to /be/, because
otherwise nginx would send API requests directly to the backend rather
than through the frontend server. In the preview, though, we do just
talk directly to the backend server, because I was having trouble
getting it to work going through nginx on localhost; I don't think
that particular problem will matter when I actually deploy to prod.