Enable parallel testing in backend
This gives around an 80% speedup in how long the backend tests take to
run.

A couple of the frontend tests break when run in parallel due to
library issues rather than my test code, so that needs to stay off for
now.

The e2e tests need to be run in sequence for now because they will
step on each other's toes if run at the same time, eg they rely on
being able to find the thread they just created on the first page of
results.